separated gradient implementations into their own files
